pyproject.toml
[tool.poetry]
name = "jtdub-com"
version = "0.1.0"
description = ""
authors = ["jtdub <james.williams@packetgeek.net>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
python-frontmatter = "^1.1.0"
markdown = "^3.7"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
bs4 = "^0.0.2"
requests = "^2.32.3"
ruff = "^0.9.2"
invoke = "^2.2.0"

[tool.pytest.ini_options]
pythonpath = ["."]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
